Move Pay By Link tests to admin folder
