+++ projects/magento/tests/backoffice/Backoffice.spec.js
@@ -2,6 +2,7 @@
 Magento doesn't allow multiple administrator sessions,
 we are running them through this file to allow serialization */
 
-import "../backoffice/MOTO.js";
-// import "../backoffice/RequiredSettingsAutoConfig.js"
-// import "../backoffice/RequiredSettingsManualConfig.js"
+import "./MOTO.js";
+import "./RequiredSettingsAutoConfig.js"
+import "./RequiredSettingsManualConfig.js"
+import "./PayByLink.js"
